requirements.txt
pact-broker-client==2.0.0
pact-python==0.8.0
Flask==0.12.2
pytest==3.2.2
requests==2.12.3
